Refactor mreg/tests.py into a package of tests.
