+++ mreg/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Base test package for mreg."""
